bce-server/requirements.txt
django
djangorestframework
djangorestframework-jwt
psycopg2
django-s3-storage
django-cors-headers
django-webpack-loader
zappa
